backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
bcrypt==3.2.2
binance-futures-connector==4.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
dateparser==1.2.1
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.12
frozenlist==1.6.0
greenlet==3.2.1
h11==0.16.0
httptools==0.6.4
idna==3.10
multidict==6.4.3
numpy==1.24.4
pandas==2.2.3
pandas-ta==0.3.14b0
passlib==1.7.4
propcache==0.3.1
psycopg2==2.9.10
pyasn1==0.4.8
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
python-binance==1.0.28
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.3.3
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
yarl==1.20.0
